pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "chiller-plant-control-data"
version = "0.1.0"
description = "CSV processing utilities for chiller plant control data."
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

